Rename `tests2.rs` to more descriptive `test_bulk_gossip.rs`
